Make StatusIndicator shareable between attendance and deliverables
