+++ src/features/Attendance/StudentStatusRow.jsx
@@ -1,13 +1,12 @@
 // Components
-import StatusIcon from "./StatusIcon/StatusIcon"
 import ProfileInfo from "../../components/ProfileInfo/ProfileInfo"
+import StatusIndicator from "../../components/StatusIndicator/StatusIndicator"
 
 const StudentStatusRow = ({ student }) => {
   // Add lookup table for status icons
   return (
     <div className="row">
-      {/* <p>{student.status}</p> */}
-      <StatusIcon status={student.status} />
+      <StatusIndicator status={student.status} />
       <ProfileInfo profile={student} />
     </div>
   )
